Strangler extraction #10: itch.io tab -> zxnu_itchio_pane.py
Move the whole optional itch.io tab block (state init, connection/auth
row, collections combo, search, table/gallery stack, item viewer with
install/uninstall/send-to-SD-card actions, cross-search registration and
the conditional tab insertion) out of MainWindow.__init__ into
build_itchio_pane(host, ...) using the proven builder-function seam:
self -> host via a tokenize NAME pass, __init__-locals injected as
keyword-only params (forwarding lambdas for the tab-spinner/badge and
cross-search helpers defined later in __init__), and the module-global
right_disk_image_explorer_content read through the _right_disk_content
getter hook. zxnu_main.py 12,869 -> 11,571 lines.

Also drop QSizePolicy from a lazy closure import the extraction exposed
as an F811 (the module-level import now serves that dialog), and add the
new module to the tests/test_pane_imports.py free-var tripwire.
